refactor amp/bento-inline-gallery
fix inline gallery deps
