+++ extensions/amp-inline-gallery/1.0/base-element.js
@@ -1,10 +1,7 @@
 import * as Preact from '#preact';
-import {CarouselContextProp} from '../../amp-base-carousel/1.0/carousel-props';
-import {BentoInlineGallery} from './component';
 import {PreactBaseElement} from '#preact/base-element';
 import {dict} from '#core/types/object';
-import {setProp} from '#core/context';
-import {useContext, useLayoutEffect} from '#preact';
+import {Component, ContextExporter, detached, props} from './element';
 
 export const TAG = 'bento-inline-gallery';
 
@@ -18,26 +15,10 @@ export class BaseElement extends PreactBaseElement {
 }
 
 /** @override */
-BaseElement['Component'] = BentoInlineGallery;
+BaseElement['Component'] = Component;
 
 /** @override */
-BaseElement['detached'] = true;
+BaseElement['detached'] = detached;
 
 /** @override */
-BaseElement['props'] = {
-  'loop': {attr: 'loop', type: 'boolean'},
-};
-
-/**
- * @param {!SelectorDef.OptionProps} props
- * @return {PreactDef.Renderable}
- */
-function ContextExporter({shimDomElement}) {
-  // Consume the `CarouselContext` produced by the `InlineGallery` component
-  // and propagate it as a context prop.
-  const context = useContext(CarouselContextProp.type);
-  useLayoutEffect(() => {
-    setProp(shimDomElement, CarouselContextProp, ContextExporter, context);
-  }, [shimDomElement, context]);
-  return <></>;
-}
+BaseElement['props'] = props;
